pyproject.toml
[project]
name = "streaming_platform_backend"
version = "0.1.0"
description = "Backend for streaming platform"
authors = [{ name="John Adeniran", email="timihack@example.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis>=2.0.1",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "black>=25.11.0",
    "fastapi[standard]>=0.121.3",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "passlib[bcrypt]>=1.7.4",
    "pre-commit>=4.4.0",
    "pydantic-settings>=2.12.0",
    "pydantic[mypy]>=2.12.4",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "types-passlib>=1.7.7.20250602",
    "types-pyjwt>=1.7.1",
    "types-redis>=4.6.0.20241004",
    "types-requests>=2.32.4.20250913",
    "uvicorn>=0.38.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
explicit_package_bases = true


[[tool.mypy.overrides]]
module = "pydantic_settings.*"
ignore_missing_imports = true

[[ tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = ["passlib.*", "jwt.*", "redis.*", "sqlmodel.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.models.*"
disable_error_code = ["call-arg"]

[tool.flake8]
max-line-length = 130
extend-ignore = ["E203", "E501"]
